requirements.txt
praw
google-generativeai
python-dotenv